controller: fix group_table and meter_table allocation
The group_table and meter_table are initialized in ovn-controller, with n_ids = 0.
Then they are re-initialized in ofctrl, with proper number of n_ids, in state S_CLEAR_FLOWS.
However, nothing prevented to start adding flows (by adding logical ports) using groups
before ofctrl reaches this state. This was causing some wrong flows (missing group in the flow).

With this patch, as soon as the feature rconn is available, i.e. before adding flows, those table
are properly re-initialized.

This issue is usually not visible in main and recent branches ci, since
"Wait for new ovn-controllers to connect to Southbound." as this was slowing down the moment when
the test started to add ports.
This was causing the following test to fail:
"ECMP static routes -- ovn-northd -- parallelization=yes -- ovn_monitor_all=yes".

Fixes: 1d6d953bf883 ("controller: Don't artificially limit group and meter IDs to 16bit.")
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>