Revert "ovn: add geneve PMTUD support"
This reverts commit 450e41e783bfa69e4f9d6c80f6bcb01147d5cfe1.

If a packet has to be tunnelled to another node and if the physical
interface used for tunnelling has lower MTU than the packet or
if there is a route exception with a lower MTU, then the geneve
kernel module generates an ICMP need frag packet.  This packet
was getting dropped since the metadata had to be swapped.
The commit did exactly that and fixed the issue.
But it has 2 issues -
  1. It introduced a regression for the scenario when an ICMP need frag
     packet generated outside of OVN has to be tunnelled and delivered
     to the destination VM/pod.  These ICMP need frag packets are now
     dropped.
  2. If the logical switches has ACLs or load balancers configured then
     these icmp need frag packets are dropped as they are not sent to
     the correct zone.

Its better to revert until we find a proper solution for the original
issue.

Reported-at: https://issues.redhat.com/browse/FDP-216

Fixes: 450e41e783bf ("ovn: add geneve PMTUD support")
Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>