Correct ethtype referencing incorrect values
documentation in OVN-NB in Logical_Switch_Port for Ethtype references
incorrect values

"Supported values: 802.11q (default), 802.11ad."

This should be 802.1q and 802.1ad.

Correct this in the code and documentation. For now I logged as a
warning that the incorrect value is used and maybe in the future the
incorrect value could be removed.

Reported-at: https://mail.openvswitch.org/pipermail/ovs-discuss/2023-December/052836.html
Fixes: 50f4ea011622 ("Support 802.11ad EthType for localnet ports")

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>