system-tests: Wait for the meter in CoPP tests
The CoPP test modifies a meter from drop=1 to drop=10, there are two
issues with this change:

1) It takes some time for this change to propagate into OvS.
2) Depending on the timing the 10 packet limit might not fit into
single pktps bucket.

To address those issues lower the pktps to 5, this has lower chance
for the bucket to overflow into the next second. Also wait for OvS
to receive the meter update and the meter stats reset.

One thing to note is that even with this change the test can still
fail with more packets being allowed through than expected, however
the chance is highly reduced. For comparison, the test was failing
on ARM in ~1/5 of runs. After this change there wasn't single failure
in a loop with 100 runs.

Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>