Add "pkt_clone_type" option to LSP.
Currently only valid value for this option is mc_unknown.
If set to mc_unknown, packets destined to the port gets cloned
to all unknown ports connected to the same Logical Switch.

Usecase:
=========
It is required to reduce packet loss when VM is being migrated to
different AZ via VXLAN tunnel. Port is configured in both AZs
on different logical switches which are sharing same IP subnet.
VTEP device is connected to the same logical switch on both AZs
and its port is set to unknown.

Signed-off-by: Naveen Yerramneni <naveen.yerramneni@nutanix.com>
Signed-off-by: Numan Siddique <numans@ovn.org>