ovn-sandbox: Rename script from ovs-sandbox to ovn-sandbox.
Update documentation.  To avoid completely duplicating the contents of the
OVS repo sandbox documentation point to it instead where applicable
(similar links already existed).

It's probably simpler to keep a (renamed) copy of the OVS ovs-sandbox
script in our repo than to use the one from the OVS submodule and build on
top.  Remove the post-split TODO item that suggested we do the latter.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>