ovn-nbctl: Document "--portrange" in the manpage.
The --portrange option was added in commit [0]. The option appears in
the help text for ovn-nbctl, but it is not documented in the manpage at
all.

This commit adds documentation to the manpage for the --portrange
option.

Reported-at: https://issues.redhat.com/browse/FDP-537
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 9e8f6f3d22e9262fdefb8347724b47cf70572dcf)