northd: Do not incrementally proccess changes for disabled LR.
When the logical router is disabled it is not populated in the datapaths
and is not internally available to I-P engine in the hmap. This could
lead to a crash during I-P processing:

To fix that make sure we process incrementally only LR that are enabled.

Reported-at: https://issues.redhat.com/browse/FDP-571
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>