controller: Nonvif related lports handling.
This patches fixes flows not properly deleted in scenarios similar to:
foo1 (hv1) - foo - R1 - join - R2 (chassis = hv2) - alice - alice1 (hv2).

When R2 is deleted, alice_R2 changed from l3gateway to patch, and, on hv1,
alice_R2 was added to related ports.
When R2 was added back (together with R2-alice and R2-join), alice_R2 changed
back from patch to l3gateway and alice_R2 remained in related_lports on hv1.

This is now fixed: a l3_gateway port is not a related_port if not for our chassis.

A test case has been modified to highlight the error, but also to
make the test easier to read (e.g avoid using same name for a port and
for a switch).

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>